Fix support for TypeScript 2.x
As the client uses Bluebird for the Promise spec we need to tell the definitions to use that version rather than the Vanilla one.

The ws.d.ts needs fixing also but as that works for both Vanilla and any spec something needs to be done custom or if something exists to cover them all.